move cli.py to __main__.py
