+++ docs/console_script_setup.rst
@@ -10,8 +10,8 @@ How It Works
 ------------
 
 If the 'command_line_interface' option is set to ['click'] during setup, cookiecutter will
-add a file 'cli.py' in the project_slug subdirectory. An entry point is added to
-setup.py that points to the main function in cli.py.
+add a file '__main__.py' in the project_slug subdirectory. An entry point is added to
+setup.py that points to the main function in `__main__`.py.
 
 Usage
 ------------
