+++ hooks/post_gen_project.py
@@ -18,9 +18,7 @@ if __name__ == '__main__':
         remove_file('tests/__init__.py')
 
     if 'no' in '{{ cookiecutter.command_line_interface|lower }}':
-        cli_file = os.path.join('{{ cookiecutter.project_slug }}', 'cli.py')
         main_file = os.path.join('{{ cookiecutter.project_slug }}', '__main__.py')
-        remove_file(cli_file)
         remove_file(main_file)
 
     if 'Not open source' == '{{ cookiecutter.open_source_license }}':
