+++ {{cookiecutter.project_slug}}/setup.py
@@ -47,7 +47,7 @@ setup(
     {%- if 'no' not in cookiecutter.command_line_interface|lower %}
     entry_points={
         'console_scripts': [
-            '{{ cookiecutter.project_slug }}={{ cookiecutter.project_slug }}.cli:main',
+            '{{ cookiecutter.project_slug }}={{ cookiecutter.project_slug }}.__main__:main',
         ],
     },
     {%- endif %}
