+++ {{cookiecutter.project_slug}}/tests/test_{{cookiecutter.project_slug}}.py
@@ -14,7 +14,7 @@ from click.testing import CliRunner
 
 from {{ cookiecutter.project_slug }} import {{ cookiecutter.project_slug }}
 {%- if cookiecutter.command_line_interface|lower == 'click' %}
-from {{ cookiecutter.project_slug }} import cli
+from {{ cookiecutter.project_slug }} import __main__
 {%- endif %}
 
 {%- if cookiecutter.use_pytest == 'y' %}
@@ -40,10 +40,10 @@ def test_content(response):
 def test_command_line_interface():
     """Test the CLI."""
     runner = CliRunner()
-    result = runner.invoke(cli.main)
+    result = runner.invoke(__main__.main)
     assert result.exit_code == 0
-    assert '{{ cookiecutter.project_slug }}.cli.main' in result.output
-    help_result = runner.invoke(cli.main, ['--help'])
+    assert '{{ cookiecutter.project_slug }}.__main__.main' in result.output
+    help_result = runner.invoke(__main__.main, ['--help'])
     assert help_result.exit_code == 0
     assert '--help  Show this message and exit.' in help_result.output
 {%- endif %}
@@ -66,10 +66,10 @@ class Test{{ cookiecutter.project_slug|title }}(unittest.TestCase):
     def test_command_line_interface(self):
         """Test the CLI."""
         runner = CliRunner()
-        result = runner.invoke(cli.main)
+        result = runner.invoke(__main__.main)
         assert result.exit_code == 0
-        assert '{{ cookiecutter.project_slug }}.cli.main' in result.output
-        help_result = runner.invoke(cli.main, ['--help'])
+        assert '{{ cookiecutter.project_slug }}.__main__.main' in result.output
+        help_result = runner.invoke(__main__.main, ['--help'])
         assert help_result.exit_code == 0
         assert '--help  Show this message and exit.' in help_result.output
 {%- endif %}
